+++ src/views/projectExplorer/index.ts
@@ -26,6 +26,7 @@ import Source from "./source";
 import SourceFile from "./sourceFile";
 import Variable from "./variable";
 import { EnvironmentManager } from "../../environmentManager";
+import SourceDirectory from "./sourceDirectory";
 
 /**
  * Represents the Project Explorer tree data provider.
@@ -41,7 +42,7 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
     ibmi?.onEvent(`connected`, () => {
       this.refresh();
 
-      currentDeploymentStorage = ibmi?.getStorage().getDeployment();
+      currentDeploymentStorage = JSON.parse(JSON.stringify(ibmi?.getStorage().getDeployment()));
     });
     ibmi?.onEvent(`deploy`, () => {
       this.refresh();
@@ -57,7 +58,7 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
         }
       }
 
-      currentDeploymentStorage = newDeploymentStorage;
+      currentDeploymentStorage = JSON.parse(JSON.stringify(newDeploymentStorage));
     });
     ibmi?.onEvent(`disconnected`, () => {
       this.refresh();
@@ -109,20 +110,31 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
           }
         }
       }),
-      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.runCompile`, async (element?: Project | Uri) => {
+      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.runCompile`, async (element?: SourceFile | SourceDirectory | Uri) => {
         let iProject: IProject | undefined;
         if (element) {
-          if (element instanceof Project) {
-            iProject = ProjectManager.get(element.workspaceFolder);
-          } else {
-            iProject = ProjectManager.getProjectFromUri(element);
+          if (element instanceof SourceFile || element instanceof SourceDirectory) {
+            // Invoked from a file or directory under the Source heading
+            element = element.sourceInfo.localUri;
           }
+
+          // Invoked from the editor or file explorer
+          iProject = ProjectManager.getProjectFromUri(element);
         } else {
-          iProject = ProjectManager.getActiveProject();
+          // Invoked from the command palette
+          let activeFileUri = window.activeTextEditor?.document.uri;
+          element = activeFileUri?.scheme === 'file' ? activeFileUri : undefined;
+
+          if (element) {
+            iProject = ProjectManager.getProjectFromUri(element);
+          } else {
+            window.showErrorMessage(l10n.t('Failed to retrieve file to compile'));
+            return;
+          }
         }
 
         if (iProject) {
-          iProject.runBuildOrCompileCommand(false);
+          iProject.runBuildOrCompileCommand(false, element);
         } else {
           window.showErrorMessage(l10n.t('Failed to retrieve project'));
         }
@@ -437,8 +449,13 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
           }
         }
       }),
-      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.setTargetLibraryForCompiles`, async (element: Uri) => {
+      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.setTargetLibraryForCompiles`, async (element: SourceDirectory | Uri) => {
         if (element) {
+          if (element instanceof SourceDirectory) {
+            // Invoked from a directory under the Source heading
+            element = element.sourceInfo.localUri;
+          }
+
           const iProject = ProjectManager.getProjectFromUri(element);
 
           if (iProject) {
@@ -514,8 +531,13 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
           }
         }
       }),
-      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.setTargetCCSIDForCompiles`, async (element: Uri) => {
+      commands.registerCommand(`vscode-ibmi-projectexplorer.projectExplorer.setTargetCCSIDForCompiles`, async (element: SourceDirectory | Uri) => {
         if (element) {
+          if (element instanceof SourceDirectory) {
+            // Invoked from a directory under the Source heading
+            element = element.sourceInfo.localUri;
+          }
+
           const iProject = ProjectManager.getProjectFromUri(element);
 
           if (iProject) {
@@ -680,34 +702,48 @@ export default class ProjectExplorer implements TreeDataProvider<ProjectExplorer
           }
         }
       }),
-      commands.registerCommand(`vscode-ibmi-projectexplorer.addToIncludePaths`, async (element: IncludePaths | any) => {
-        if (element instanceof IncludePaths) {
-          const iProject = ProjectManager.get(element.workspaceFolder);
+      commands.registerCommand(`vscode-ibmi-projectexplorer.addToIncludePaths`, async (element: IncludePaths | SourceDirectory | Uri | any) => {
+        if (element) {
+          let iProject: IProject | undefined;
+          let includePath: string | undefined;
+          if (element instanceof IncludePaths) {
+            // Invoked from Include Paths heading
+            iProject = ProjectManager.get(element.workspaceFolder);
 
-          if (iProject) {
-            const includePath = await window.showInputBox({
-              prompt: l10n.t('Enter include path'),
-              placeHolder: l10n.t('Include path')
-            });
+            if (iProject) {
+              includePath = await window.showInputBox({
+                prompt: l10n.t('Enter include path'),
+                placeHolder: l10n.t('Include path')
+              });
+            }
+          } else if (element instanceof SourceDirectory) {
+            // Invoked from a directory under the Source heading
+            includePath = element.sourceInfo.localUri.path;
+
+            iProject = ProjectManager.get(element.workspaceFolder);
+          } else if (element instanceof Uri) {
+            // Invoked from the file explorer
+            includePath = element.path;
+
+            iProject = ProjectManager.getProjectFromUri(element);
+          } else {
+            // Invoked from the Code4i IFS browser
+            includePath = element.path;
 
             if (includePath) {
-              await iProject.addToIncludePaths(includePath);
+              iProject = ProjectManager.getActiveProject();
+            } else {
+              window.showErrorMessage(l10n.t('Failed to retrieve path to directory'));
+              return;
             }
-          } else {
-            window.showErrorMessage(l10n.t('Failed to retrieve project'));
           }
-        } else {
-          const includePath = element.path;
 
           if (includePath) {
-            const iProject = ProjectManager.getActiveProject();
             if (iProject) {
               await iProject.addToIncludePaths(includePath);
             } else {
               window.showErrorMessage(l10n.t('Failed to retrieve project'));
             }
-          } else {
-            window.showErrorMessage(l10n.t('Failed to retrieve path to directory'));
           }
         }
       }),